Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-float a,b,c;
-scanf("%f %f %f",&a,&b,&c);
-printf("Average: %f", a+b+c/2);
+int a,b,c;
+scanf("%d %d %d",&a,&b,&c);
+printf("Average: %d", a+b+c/2);
     return 0;
 }
